:recycle: [open-formulieren/open-forms#5133] Refactor step summary progress
Refactored to perform the index/progress calculation in the component,
as this component/functionality is shared between the legacy and new
form step implementation, allowing us to benefit from typescript in
the process.

In the FormStep/index.js, the component is now moved inside the
existing guard if nothing is loading, which should not make any
functional difference.

The check whether something needs to be rendered or not is moved into
the component for better self-containment.