:recycle: [open-formulieren/open-forms#5133] Refactor previous/next/last step utilities
Instead of doing multiple calculations over and over and keeping temp
variables around that clutter the react components, use a utility data
container with properties/methods to derive this information.

This centralizes the helpers and makes them easier to debug, while
adding type information and moving the utility code in a semantically
more correct location.