:sparkles: [open-formulieren/open-forms#5378] Start drawing on map click, when only 1 interaction is allowed
When only 1 interaction is allowed, it makes little sense to pick a shape (as there is only 1 shape available). So to reduce the amount of user clicks, when only 1 interaction is active you can start drawing by simple clicking the map.

To prevent visual noise, the "draw" buttons are only shown when there are multiple available interactions.