:white_check_mark: [open-formulieren/open-forms#5191] Added map component tests for new "use current location" control.
The tests cover the situations of:
1. 'permission already granted'
2. 'permission already denied'
3. 'permission gets denied by user'

To mock the permission API and geolocation, a new 'mockGeolocationDecorator' decorator was added. This decorator also exposes a new argument, `_updateGeolocationPermission`, used to simulate permission changes during a storybook test. This allows us to showcase the situation where the user gives or denies permission while using the form.