:white_check_mark: [open-formulieren/open-forms#5191] Added map component tests for new "use current location" control.
The tests cover the situations of:
1. 'permission already granted'
2. 'permission already denied'
3. 'permission gets denied by user'

To mock the permission API and geolocation, a new 'setupGeolocationMock' function was added. Using the 'withGeolocationMocking' decorator, this mock function is initialized into the storybook global configuration. Using the parameters of stories, you can control the geolocation permission and lat/lng for that test. By default, the permission state 'granted' and the lat/lng of the current Maykin office are used. By using the new `parameters.geolocation.updatePermission` parameter in a storybook test, you can trigger/simulate dynamic permission changes, testing the expected behaviour of permission listeners.