:rotating_light: Enable import linting
The goal here is to have a linter complain about type-only
imports, requiring the 'type' keyword to be present.

Personally, I'd prefer to have imports statements that
are only type imports have a single top-level type keyword,
and imports that are a mix of runtime and type imports
use the inline type keyword, but that doesn't seem possible
with the available import rules (ts-eslint or import-eslint
plugins), which brings us to separate import statements
for types and runtime information. I can live with it.