:bug: Only make API call if we have a valid date
If the user types in a date, it is not valid until
all parts have been entered. Sending invalid dates
to the backend triggers validation errors, which crash
the appointment step.

Now the API call is only made when a valid date is
provided, and the mocks are updated to simulate the
backend behaviour.