:art: [open-formulieren/open-forms#5191] Extended map location permission change handling
Previously the location control would only update the control properties when the permission would change for 'positive' (granted or prompt) to 'negative' (denied), but a permission change for negative to positive is ofcourse also possible.

First, by moving the 'positive' location control property assignment to an "enable" function, we can easily toggle between the enabled and disabled states.

Second, by handling all permission changes in the `this._getPermission` callback, we can handle the permission change from negative to positive.

These two changes put together allows us to respond to any permission changes. And, because the `this._getPermission` callback is also called for the initial state, the initial render will always show the current correct UI.