:sparkles: [open-formulieren/open-forms#5576] Add WFS layers to overlays control
Leaflet doesn't support WFS layers out of the box. Contrary to WMS layers, WFS need to be rendered as geojson objects.

Our custom WFS tile layer component fetches the WFS data when moving the map, renders it as geojson into a LayerGroup, which is then placed in the map. By rendering it into an actual LayerGroup component, we can still toggle the layer using the `LeafletLayersControl.Overlay`.

PDOK, a popular Dutch provider for map layer data, doesn't support multiple Feature types when requesting the output to be json. To still support one overlay with multiple feature types, we have to spread the WFS data request to multiple individual requests, and merge the results into one LayerGroup.