:package: [#445] Re-organize dist/package structure
The folder layout in dist/ is being updated to make it easier
to distribute the TS types in the NPM package, which are
arguably the people benefiting from them the most.

This requires some inversion of the current structure,
where the bundles now land directly in dist/ and the
ESM module build ends up in dist/esm, the future
situation puts the ESM modules in dist/ and the full
bundles in a subdirectory dist/bundles/.

With the package export maps we can keep the same
import interfaces and only point to the new paths,
maintaining backwards compatibility, and at the nginx
container image level we can provide a backwards
compatible setup through symlinks.

Additionally, some dead exports have been cleaned
up/removed, and the package README updated. Some
original ambitions to provide a pluggable interface
saw no adoption at all, so that's been removed now,
since the code for this has been removed a long time
ago already anyway.