[#445] :label: Remaining TS conversion
From PR #891

* :label: [#445] Convert remaining stories/tests to TS

* :recycle: [#445] Move login URL utilities into more appropriate location

This is now colocated closer to the code dealing with login URLs or
login options for the form, and some helpers are put in the same
module as where they're used because they are only used in a single
place.

* :fire: [#445] Delete obsoleted stories

The components (and stories) here have now live in the formio-renderer
package. We only kept these stories temporarily for the visual
regression testing, but these visual regression tests now live
in the formio-renderer Chromatic itself, so they can be removed
here.

* :recycle: [#445] Update select field stories

Only document the difference between static/dynamic options and
remove the other stories, as those are covered in the formio-renderer
package.

* :label: [#445] Convert legacy cosign to TS

... and cleaned up the implementation.

There were quite some props sprinkled around that were
effectively not used anymore in a *long* time, after
digging through the commit history for 4 file renames
and ending up back in 2022-ish.

* :fire: Delete obsoleted stories

These are either obvious or have been moved into the formio-renderer package.

* :label: [#445] Convert test to TS

* :art: [#445] Update imports to components that are TS

* :recycle: [#445] Move config decorator to global storybook config

The config decorator is nearly almost required, and adding it
as a global decorator makes sense because the entire app
is contained in a config context.

The same (story) parameters can still be used to influence
configuration behaviour, and the moved decorator is now
properly typed.

* :recycle: [#445] Incorporate the analytics tools decorator into config decorator

This config is nearly always available (except for in appointment
flows) as it's rendered by the Form component. Configuration still
uses the same parameters.

* :recycle: [#445] Refactor LayoutDecorator and move to decorators module

* :truck: [#445] Move the withCard decorator

* :recycle: [#445] Move the withForm decorator

And refactored it to not grab the form from
args at all anymore, but instead only consider
the story parameters.

* :recycle: [#445] Properly convert the submission poll info and modal decorators to TS

* :recycle: [#445] Refactor the LiteralsDecorator

* Convert to TS
* Move to decorators module
* Grab config from params instead of args

* :recycle: [#445] Move withFormik decorator

* :label: [#445] Convert Storybook preview config to TS

* :label: [#445] Update imports in stories

* :coffin: [#445] Delete dead prop types code

* :label: [#445] Convert remaining hooks/tests to TS

* :label: [#445] Convert cache utilities to Typescript

* :rotating_light: [#445] Configure ESLint to disallow JS instead of TS

JS is only allowed when dealing with formio.js, but in all other places,
developers *must* use typescript.

* :art: [#445] Normalize imports in typescript files

All absolute imports should start with @/ to indicate
they're absolute, and this makes the import ordering
configuration easier.

Legacy JS modules are untouched.