:zap: [open-formulieren/open-forms#5576] Reuse WFS overlay data and limit new data requests
Instead of fetching new WFS overlay data on every map movement, we only fetch data for the "new" map tiles. This limits the amount of WFS data requests significantly.
As added bonus, this also reduces map layer rebuilding/re-rendering, providing a nicer user experience.

In the LeafletMapLayersControl a `tileCacheRef` is added. This is a simple set of all the map tiles that were fetched for the WFS overlay. When the map is moved, only the visible tiles that are not cached will be fetched.

After fetching WFS overlay data, we no longer clear/remove the previous WFS data. Instead, the new overlay data is simply added to the current LayerGroup.

When the zoom level changes the tile cache and LayerGroup are cleared, and new WFS data is requested. This ensures that the shown features remain accurate.

Additionally, we now only fetch new map data if the WFS layer is actually shown. If the layer is visually hidden, there is no need to fetch new visual data.