:whale: Fix creating nested static symlinks
Both the bundles and the ESM builds produce a static folder
containing the static assets, which are identical for each
flavour.

We don't need to create a symlink here then for backwards
compatibility. The symlink currently causes problems
with django's collectstatic command because it's not going
anywhere.

Additionally, the symlinks to the source map URLs were
missing.