:whale: Fix Docker image build for a specific version
The goal is to create a dist/VERSION subdirectory containing the assets for
a particular versioned build (whereas latest just uses dist/ as target
directory).

This worked fine for a long time, but the vite-dts plugin looks at the
package.json exports for the types, which points to dist/bundles/index.d.ts,
as a relative path inside the package. This configuration is correct,
for both a 'latest' build and a particular versioned build, as the npm
package itself is already versioned and uses fixed output paths.

However, the docker image build was creating this subdir. For legacy
reasons (all the way back to CRA/webpack) there was a need for that,
but it no longer exists since we bake the SDK assets into the backend
image anyway. For local development, there's also not much added value
to this feature, so the vite config is simplified to always just output
to dist/ and use fixed paths, whatever the version is.

We now handle the subdirectory (backwards compatibility and cache
busting mechanisms in the docker image) exclusively in the docker
build phase.