:wrench: Inline the formio-renderer in vitest
This helps ensure that itself and its dependencies are processed from
CommonJS and ESM interop exports are added, which actually makes the
tests pass.

Other options that were tried to resolve this:

* use react-use/esm/useAsync imports in the renderer, but that
  breaks on module resolution of relative import paths without
  extension (and it's fragile)
* resolve.alias approach, which requires special care to not
  mess up react-use/esm/\* imports, and still doesn't work
* inlining only react-use - doesn't seem to be picked up

I'm at a loss here - vitest/node seem hellbent on externalizing
this dependency *and* loading it as is in CommonJS, and then
also completely ignore the module.exports that are present
in the code. If this makes it work for now, fine...

More details here: https://github.com/vitest-dev/vitest/discussions/4233