:lipstick: Fix error styling regression
The renderer includes styles to preserve linebreaks in validation error
messages, however, this gets applied to the entire form field component
that's invalid and messes with the newlines produced by the formio
EJS templates.

So, in the formio context, we revert those styles (which are not a
problem with React) and apply them to more specific template nodes.