:bug: [open-formulieren/open-forms#5727] Using map overlays in summary
Ensuring map overlays are also shown in the form summary.

To keep the form summary UI bloat to a minimum, the map component should contain as few controls as possible. To that end, the controls for the overlays will not be shown in the form summary. (all the overlays will always be shown)

This also mimics the behaviour of the submission email, where the map shows all its overlays.

Backport-Of: open-formulieren/open-forms#5727