[open-formulieren/open-forms#5548] Updated add-button to secondary action button
This PR handles only the following in the SDK. Some other changes are
needed in the new formio renderer (for example in the edit grid
component, which is used in the appointments). Discussed the email
verification and concluded that the button will stay as is because the
secondary would make it hard to find in some themes.

- EditGrid
- Partners
- Children
- Formio's multiple True