:sparkles: [#5821] Add customerProfile component to form summary
Added valueDisplay component for the form component CustomerProfile. This will be used to display profile data in form-step and confirmation summaries.

The value will be displayed as a simple list of addresses, with an additional suffix for future preferred addresses. This is similar to the display used for the submission pdf.