:children_crossing: [#5817] Implement profile component updatePreferencesModalEnabled
When the profile component is used by non-authenticated users, we should hide the "update preferences" button. In order to update profile preferences we need an authentication identifier.

Currently we only support BSN authentication. This can however not be enforced in the JS code. We make it clear in the configuration and documentation that the Profile component currently only works for BSN authentication.